working files
This commit hacks support for each of the distance/color overlays
for the relevant Galactic properties by hacking the menu and renaming a bunch
of files. **THE DATA IS ALL IDENTICAL AND ONLY MEAN AS A PLACEHOLDER.**
Also modifies some of the UI to better match units, although the color
schemes have not been adjusted.